requirements.txt
psutil
scikit-learn
numpy
python-dotenv
requests
rich
aptos-sdk
